LPS-106397 Move base class to it's own package.
