+++ modules/core/petra/petra-sql-dsl-api/src/main/resources/com/liferay/petra/sql/dsl/base/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
