+++ modules/core/petra/petra-sql-dsl-spi/src/main/java/com/liferay/petra/sql/dsl/spi/factory/DefaultColumnFactory.java
@@ -14,8 +14,8 @@
 
 package com.liferay.petra.sql.dsl.spi.factory;
 
-import com.liferay.petra.sql.dsl.BaseTable;
 import com.liferay.petra.sql.dsl.Column;
+import com.liferay.petra.sql.dsl.base.BaseTable;
 import com.liferay.petra.sql.dsl.factory.ColumnFactory;
 import com.liferay.petra.sql.dsl.spi.DefaultColumn;
